Simplified 2D topology example using argparse
